Split into separate MethodCallHandler class and resolve issue
